requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-socketio==5.10.0
python-multipart==0.0.6
python-dotenv==1.0.0
numpy==1.26.4
opencv-python==4.8.1.78
mediapipe==0.10.8
torch==2.1.1
torchvision==0.16.1
pydantic==2.5.2 